web/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
PyPDF2>=3.0.0
python-docx>=1.1.0
python-pptx>=0.6.0
pytest>=7.4.0
httpx>=0.25.0

